Use a buffer inside count_words
* src/extractwords.c (count_words): Use a fixed length buffer to store
the current word string and pass this to add_word_to_tree.  If the
buffer was to small start growing the string on the obstack and pass
NULL to add_word_to_tree.
(add_word_to_tree): Change signature and update caller.  If the supplied
buffer is not NULL use this as string and assume that no growing string
is on top of the obstack.
* tests/CMakeLists.txt (test_word_frequency): Update compile features.
* tests/test_word_frequency.c: Add tests.